+++ wayland-protocols/build.rs
@@ -64,8 +64,11 @@ static WLR_UNSTABLE_PROTOCOLS: &[VersionedProtocol] = &[
     ("wlr-virtual-pointer", &[("v1", &[])]),
 ];
 
-static MISC_PROTOCOLS: &[StableProtocol] =
-    &[("gtk-primary-selection", &[]), ("server-decoration", &[])];
+static MISC_PROTOCOLS: &[StableProtocol] = &[
+    ("gtk-primary-selection", &[]),
+    ("input-method-unstable-v2", &[]),
+    ("server-decoration", &[])
+];
 
 fn generate_protocol(
     name: &str,